Add files via upload
The 2021 Behavioral Risk Factor Surveillance System (BRFSS) is a state-based telephone survey conducted by the Centers for Disease Control in collaboration with state health departments; initiated in 1984, it collects data on health behaviors and chronic conditions with a target population of non-institutionalized persons above the age of 18 years in the United States. Conducted annually since 1989, BRFSS has grown to include all 50 United States, the District of Columbia, and U.S. territories.
The objective of the BRFSS is to “collect uniform state-specific data on health risk behaviors, chronic diseases and conditions, access to health care, and use of preventive health services related to the leading causes of death and disability in the United States”. The data are used by state health departments for various purposes, including for proposing legislation for health initiatives and measuring progress toward state health objectives. Ultimately, the BRFSS aims to monitor trends in health behaviors across the United States to inform the development of public health programs and policies.
The BRFSS sampling design consists of three stages: the selection of primary sampling units (or PSU, which are typically counties or groups of counties), the selection of households within each PSU using the RDD sampling method, and the selection of one adult aged 18 years or older from each household. To account for the possibility of nonresponse bias, the survey weights the data using post-stratification adjustment and raking ratio estimation techniques. 
Using a random-digit-dialing (RDD) sampling method, the BRFSS selects households to ask questions on a wide range of health behaviors, chronic conditions, and preventive health practices. To ensure consistency in the questions asked across all states, the survey uses a (1) core set of questions, followed by (2) optional BRFSS modules, and ending with (3) state-specific questions (including questions that address state-specific health issues). The 2021 BRFSS included 438,693 observations from 49 states (in 2021, Florida was not able to collect enough data to meet the minimum requirements for inclusion in the BRFSS), the District of Columbia, and three U.S. territories (Guam, Puerto Rico, and the Virgin Islands). The reported response rate for the 2021 BRFSS, 49.7%, was similar to response rates in previous years.
The data used in this analysis were collected starting from the months of January through December of 2021, with interviews averaging at 22 minutes. Surveys were conducted in both English and Spanish, with Spanish interviews making up 7 percent of all interviews. Due to the COVID-19 pandemic, some states experienced higher levels of nonresponse than others in 2021.